parkes ground software 0.5.23
The Downlink Update - fixed downlinking and comms by expanding timeout detection and handling (still work to do) and introducing a brand new external module called pave_comms, which is a common module that can be accessed by all three software systems to create a cohesive standard that is common across systems. It seemed silly that the three computers should all communicate in different ways, and for different communication standards even within a single program - I've wanted to do this for a while. 

Should also help to cut down the length of parkes_master.py.